reorganize api_handler methods
