+++ src/screens/AddEventScreen.js
@@ -14,7 +14,7 @@ import {
     KeyboardView,
     SubmitButton
 } from '../components';
-import { addEvent } from '../utils/api_handler';
+import { addEvent } from '../utils/api_handler/events';
 
 const validationSchema = Yup.object().shape({
     eventTitle: Yup.string()
