+++ src/screens/ScheduleScreen.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { FlatList, TouchableOpacity } from 'react-native';
 
 import { Screen, EventCard } from '../components';
-import { getRecentEvents, getNextRecentEvents } from '../utils/api_handler';
+import { getRecentEvents, getNextRecentEvents } from '../utils/api_handler/events';
 import { storeJsonData, getJsonData } from '../utils/cache_handler';
 
 // Flatlist of upcoming events
